feat: create new model for acne seveity (#5)

+++ set_up_ml_mode.py
@@ -2,47 +2,45 @@ import requests
 from tqdm import tqdm
 import os 
 
-# URL of your shared Dropbox links (make sure they are modified for direct download)
+# URL of your shared Dropbox links
 yolov9_for_acne_detection_model = 'https://www.dropbox.com/scl/fi/s4tat0gyr3wiiv9nzmbua/best.pt?rlkey=8lygobguxvubacr0v6520oms4&st=w2mdh0o7&dl=1'
-mobileViT_for_acne_severity_model = 'https://www.dropbox.com/scl/fi/vr4vzwjuoom40ooviedsg/best_model.keras?rlkey=sqd5rjtdum31o4bkzhp3du8le&st=9ydkmfwr&dl=1'
 convnext_for_skin_classification_model = 'https://www.dropbox.com/scl/fi/9eds3jnzwf68rss2z7l6l/model_convnext_88.pth?rlkey=ksgbscw63yuc8w4fzetzqqahs&e=1&st=mwz3i9qw&dl=1'
+efficientnet_b3_mask = 'https://www.dropbox.com/scl/fi/0iasjhktjs2pq8funn6ob/best.pth?rlkey=lxowps4aihef96y5z25d955fc&st=7o3qfbfw&dl=1'
+unetpp = 'https://www.dropbox.com/scl/fi/o5jg8kkh3dcb0zgqy435z/best_model.keras?rlkey=zq49w1t9bir2y3wit4al37egi&st=u8dbxcd8&dl=1'
 
-# Local paths where you want to save the models
+# Local paths
 yolov9_for_acne_detection_model_path = './app/ml_models/yolo/weights/best.pt'
-mobileViT_for_acne_severity_model_path = './app/ml_models/mobileViT/weights/best_model.keras'
 convnext_for_skin_classification_model_path = './app/ml_models/convnext/model_convnext_88.pth'
+efficientnet_b3_mask_path = './app/ml_models/efficientnet_b3_mask/weights/best.pth'
+unetpp_path = './app/ml_models/unetpp/weights/best_model.keras'
 
-# Function to download a model from Dropbox with progress bar
 def download_model(url, save_path):
-    # Tạo thư mục nếu chưa tồn tại
     os.makedirs(os.path.dirname(save_path), exist_ok=True)
+
+    # Kiểm tra nếu file đã tồn tại thì bỏ qua
+    if os.path.exists(save_path):
+        print(f"File already exists at {save_path}, skipping download.")
+        return
 
     response = requests.get(url, stream=True)
 
-    # Check if the request was successful
     if response.status_code == 200:
-        # Get the total file size
         total_size_in_bytes = int(response.headers.get('content-length', 0))
 
-        # Open the file to save the model
         with open(save_path, 'wb') as f:
-            # Initialize the progress bar
-            with tqdm(total=total_size_in_bytes, unit='B', unit_scale=True, desc=f"Downloading {save_path}") as bar:
-                # Write the content to file in chunks
+            with tqdm(total=total_size_in_bytes, unit='B', unit_scale=True, desc=f"Downloading {os.path.basename(save_path)}") as bar:
                 for data in response.iter_content(chunk_size=1024):
                     f.write(data)
-                    bar.update(len(data))  # Update the progress bar
-        print(f"Model successfully downloaded and saved as {save_path}")
+                    bar.update(len(data))
+        print(f"Downloaded and saved: {save_path}")
     else:
-        print(f"Failed to download the model. Status code: {response.status_code}")
-
+        print(f"Failed to download {url}. Status code: {response.status_code}")
 
 def main():
-    # Tải cả hai model với thanh tiến trình
     download_model(yolov9_for_acne_detection_model, yolov9_for_acne_detection_model_path)
-    download_model(mobileViT_for_acne_severity_model, mobileViT_for_acne_severity_model_path)
     download_model(convnext_for_skin_classification_model, convnext_for_skin_classification_model_path)
-
+    download_model(efficientnet_b3_mask, efficientnet_b3_mask_path)
+    download_model(unetpp, unetpp_path)
 
 if __name__ == "__main__":
     main()
